Improve `Message` component
Save sid as message param in state when running `SEND_MESSAGE_SUCCESS`
reducer--the sid is a helpful unique identifier of a given twilio
message. For the same reason, use Sid as key when rendering message
on page.

Format message date to be human-readable, as this will be clearer to
users than the number of seconds since the unix epoch.